refactor: wave10 peels — agents types, mission/messaging APIs, workflow-definitions
Behavior-preserving package organization after wave 9 (#2274):
- types/agents.ts: agent permissions, entity, ratings, reflections, heartbeat runs
- app/api: missions, messaging, plugins-and-skills, todo, insights, system-panel
- rename remaining-ops-8 → task-store/workflow-definitions.ts
Ratchet types.ts and legacy.ts line-count ceilings downward.
