+++ packages/core/src/task-store/workflow-definitions.ts
@@ -1,5 +1,7 @@
 /**
- * remaining-ops-8 operations.
+ * FNXC:CodeOrganization 2026-07-18-14:00:
+ * Domain rename from remaining-ops-8: workflow definition CRUD, selection
+ * materialization, plugin gate verdicts, CLI autonomy approvals, and related store ops.
  *
  * FNXC:StoreModularization 2026-06-25-00:00:
  * Extracted from the monolithic packages/core/src/store.ts as a pure
