+++ scripts/line-count-baseline.json
@@ -8,17 +8,17 @@
   "packages/cli/src/commands/task.ts": 2366,
   "packages/cli/src/extension.ts": 5687,
   "packages/core/src/agent-store.ts": 3484,
-  "packages/core/src/async-mission-store-queries.ts": 2067,
+  "packages/core/src/async-mission-store-queries.ts": 2093,
   "packages/core/src/central-core.ts": 4486,
-  "packages/core/src/index.gate.ts": 2217,
-  "packages/core/src/index.ts": 2501,
+  "packages/core/src/index.gate.ts": 2228,
+  "packages/core/src/index.ts": 2512,
   "packages/core/src/mission-store.ts": 4369,
-  "packages/core/src/postgres/schema/project.ts": 2150,
+  "packages/core/src/postgres/schema/project.ts": 2155,
   "packages/core/src/postgres/sqlite-migrator.ts": 2112,
   "packages/core/src/store.ts": 2746,
-  "packages/core/src/types.ts": 7101,
+  "packages/core/src/types.ts": 6165,
   "packages/dashboard/app/App.tsx": 2059,
-  "packages/dashboard/app/api/legacy.ts": 10273,
+  "packages/dashboard/app/api/legacy.ts": 8913,
   "packages/dashboard/app/components/AgentDetailView.tsx": 5567,
   "packages/dashboard/app/components/AgentsView.tsx": 2254,
   "packages/dashboard/app/components/ChatView.tsx": 4020,
@@ -31,7 +31,7 @@
   "packages/dashboard/app/components/QuickEntryBox.tsx": 2562,
   "packages/dashboard/app/components/SettingsModal.tsx": 5002,
   "packages/dashboard/app/components/TaskCard.tsx": 3818,
-  "packages/dashboard/app/components/TaskDetailModal.tsx": 6506,
+  "packages/dashboard/app/components/TaskDetailModal.tsx": 6509,
   "packages/dashboard/app/components/TerminalModal.tsx": 3322,
   "packages/dashboard/app/components/WorkflowNodeEditor.tsx": 5685,
   "packages/dashboard/app/components/__tests__/AgentsView.test.tsx": 2854,
@@ -50,7 +50,7 @@
   "packages/dashboard/app/components/__tests__/TaskCard.test.tsx": 7548,
   "packages/dashboard/app/components/__tests__/TaskChatTab.test.tsx": 3141,
   "packages/dashboard/app/components/__tests__/TaskDetailModal.inline-editing-and-integrations.test.tsx": 3438,
-  "packages/dashboard/app/components/__tests__/TaskDetailModal.rendering.test.tsx": 2801,
+  "packages/dashboard/app/components/__tests__/TaskDetailModal.rendering.test.tsx": 2821,
   "packages/dashboard/app/components/__tests__/TaskDetailModal.responsive-and-dependencies.test.tsx": 2042,
   "packages/dashboard/app/components/__tests__/TerminalModal.test.tsx": 9176,
   "packages/dashboard/app/components/__tests__/UsageIndicator.test.tsx": 2931,
@@ -64,9 +64,9 @@
   "packages/dashboard/src/__tests__/github.test.ts": 2453,
   "packages/dashboard/src/__tests__/routes-auth.test.ts": 5510,
   "packages/dashboard/src/__tests__/routes-automation.test.ts": 3089,
-  "packages/dashboard/src/__tests__/routes-github.test.ts": 3792,
-  "packages/dashboard/src/__tests__/routes-planning.test.ts": 5265,
-  "packages/dashboard/src/__tests__/routes-tasks.test.ts": 2949,
+  "packages/dashboard/src/__tests__/routes-github.test.ts": 3798,
+  "packages/dashboard/src/__tests__/routes-planning.test.ts": 5271,
+  "packages/dashboard/src/__tests__/routes-tasks.test.ts": 3024,
   "packages/dashboard/src/__tests__/usage.test.ts": 5261,
   "packages/dashboard/src/chat.ts": 3069,
   "packages/dashboard/src/github.ts": 4877,
@@ -81,7 +81,7 @@
   "packages/engine/src/__tests__/cron-runner.test.ts": 2334,
   "packages/engine/src/__tests__/executor-prompt.test.ts": 2716,
   "packages/engine/src/__tests__/executor-worktree.test.ts": 2723,
-  "packages/engine/src/__tests__/heartbeat-executor.test.ts": 4455,
+  "packages/engine/src/__tests__/heartbeat-executor.test.ts": 4462,
   "packages/engine/src/__tests__/heartbeat-scheduler.test.ts": 3327,
   "packages/engine/src/__tests__/merger-merge-lifecycle.test.ts": 3697,
   "packages/engine/src/__tests__/merger-verification.test.ts": 3177,
@@ -91,10 +91,10 @@
   "packages/engine/src/__tests__/restart.integration.test.ts": 2022,
   "packages/engine/src/__tests__/self-healing.test.ts": 11692,
   "packages/engine/src/__tests__/step-session-executor.test.ts": 3226,
-  "packages/engine/src/__tests__/triage.test.ts": 7392,
+  "packages/engine/src/__tests__/triage.test.ts": 7398,
   "packages/engine/src/agent-heartbeat.ts": 5370,
-  "packages/engine/src/agent-tools.ts": 5230,
-  "packages/engine/src/executor.ts": 20110,
+  "packages/engine/src/agent-tools.ts": 5258,
+  "packages/engine/src/executor.ts": 20123,
   "packages/engine/src/merger-ai.ts": 2401,
   "packages/engine/src/merger.ts": 11208,
   "packages/engine/src/pi.ts": 2814,
